Move split constants to constant class
