+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/codegen/JvmConstants.java
@@ -384,6 +384,19 @@ public class JvmConstants {
     public static final String TYPE_NOT_SUPPORTED_MESSAGE = "JVM generation is not supported for type ";
 
     public static final int MAX_MEMBERS_PER_METHOD = 100;
+    public static final int MAX_TYPES_PER_METHOD = 100;
+    public static final int MAX_FIELDS_PER_SPLIT_METHOD = 500;
+    public static final int MAX_MODULES_PER_METHOD = 100;
+    public static final int MAX_CALLS_PER_CLIENT_METHOD = 100;
+    public static final int MAX_CONSTANTS_PER_METHOD = 100;
+    /*
+    MAX_STRINGS_PER_METHOD is calculated as below.
+    No of instructions required for create ballerina string constant object = 12
+    Java method limit = 64000
+    Max strings constant initializations per method = 64000/12 -> 5000
+    */
+    public static final int MAX_STRINGS_PER_METHOD = 5000;
+
 
     private JvmConstants() {
     }
